Modified the rubric text
Modified the rubric text to reflect lessons learned over
the past 2 semesters of grading. Please review and use this
as a starting point, not an ending point.

Note that one of purely formatting changes is to start each
sentence on its own line. This will allow us to track changes
better when reviewing the different versions of the document.